inference_server/requirements.txt
torch>=2.1.0
transformers>=4.45.0
accelerate>=0.34.0
bitsandbytes>=0.41.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
Pillow>=10.0.0
loguru>=0.7.0
pydantic>=2.5.0
